Initial description I guess
Still no bookmarks, because I have so many that I will have to take a
day to organize them in a way that makes sense here.  (I think the
internal ID's that my browser has are at 1.5k.  Though that includes
temporary ones when I'm working on something (and folders).)

Signed-off-by: anderium <33520919+anderium@users.noreply.github.com>